requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.3
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
toml==0.10.2
tornado==6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.7.2
